+++ application/models/Bentobox_model.php
@@ -1,34 +1,43 @@
 <?php
 
-class Bentobox_model extends CI_Model
+class Bentobox_model extends CSV_Model
 {
-    public function __construct()
+    function __construct()
     {
-        parent::__construct();
+        parent::__construct("bentoBoxes.csv");
     }
 
     public function getBentoBoxes() {
-        return [
-           [
-               "code" => 1,
-               "name" => "My Bento Box",
-               "accessories" => [
-                   1 => 1
-               ]
-           ]
-        ];
+        $bentoBoxes = [];
+        foreach($this->all() as $e) {
+            $data = $e->data;
+            $accessories = [];
+
+            foreach(explode("A", $data["accessories"]) as $a) {
+                if(strlen($a) == 0)
+                    continue;
+
+                array_push($accessories, (int)$a);
+            }
+
+            $data["accessories"] = $accessories;
+            $data["code"] = (int)$data["code"];
+            array_push($bentoBoxes, $data);
+        }
+
+        return $bentoBoxes;
     }
 
     public function getBentoBox($code) {
-        return [
-            "code" => $code,
-            "name" => "My Bento Box",
-            "accessories" => [
-                "1" => 1
-            ]
-        ];
+        foreach($this->getBentoBoxes() as $b) {
+            if($b["code"] == $code)
+                return $b;
+        }
+
+        return null;
     }
 
     public function saveBentoBox($bentoBox) {
+        //To be implemented later.
     }
 }